+++ packages/xprog-prensa/primitives/index.ts
@@ -1,5 +1,6 @@
 export * from './Accordion/index'
 export * from './Block/index'
 export * from './Button/index'
+export * from './Drawer/index'
 export * from './IconButton/index'
 export * from './Typography/index'