+++ packages/xprog-prensa/primitives/types.ts
@@ -1,5 +1,6 @@
 export * from './Accordion/types'
 export * from './Block/props'
 export * from './Button/props'
+export * from './Drawer/types'
 export * from './IconButton/props'
 export * from './Typography/types'